requirements.txt
# Core dependencies
pandas==2.1.4

# LangChain ecosystem
langchain==0.1.0
langchain-community==0.0.13
langchain-google-genai==0.0.11

# Vector store
chromadb==0.4.22

# Embeddings
sentence-transformers==2.3.1

# Environment management
python-dotenv==1.0.0

# LangSmith tracing
langsmith==0.0.77